+++ src/validations/ADR45.ts
@@ -1,5 +1,5 @@
 import { Entity } from '@dcl/schemas'
-import { ContentValidatorComponents, DeploymentToValidate, OK, validationFailed } from '../types'
+import { ContentValidatorComponents, DeploymentToValidate, OK, validationFailed, ValidationResponse } from '../types'
 import { ADR_45_TIMESTAMP } from './timestamps'
 
 const entityIsNotVersion3 = (entity: Entity) => entity.version !== 'v3'
@@ -10,7 +10,7 @@ const entityWasDeployedAfterADR45 = (entity: Entity) => entity.timestamp > ADR_4
  * Validate that entity meets ADR-45 validations
  * @public
  */
-export function adr45(components: ContentValidatorComponents, deployment: DeploymentToValidate) {
+export async function adr45ValidateFn(deployment: DeploymentToValidate): Promise<ValidationResponse> {
   const { entity } = deployment
 
   if (entityIsNotVersion3(entity) && entityWasDeployedAfterADR45(entity))
