+++ src/validations/profile.ts
@@ -1,13 +1,20 @@
 import { Avatar, EntityType } from '@dcl/schemas'
 import { parseUrn } from '@dcl/urn-resolver'
 import sharp from 'sharp'
-import { ContentValidatorComponents, DeploymentToValidate, OK, ValidateFn, validationFailed } from '../types'
+import {
+  ContentValidatorComponents,
+  DeploymentToValidate,
+  OK,
+  ValidateFn,
+  validationFailed,
+  ValidationResponse,
+} from '../types'
 import {
   validateAfterADR45,
   validateAfterADR74,
   validateAfterADR75,
   validateAll,
-  validateIfTypeMatches
+  validateIfTypeMatches,
 } from './validations'
 
 /** Validate that given profile deployment includes a face256 thumbnail with valid size */
@@ -15,8 +22,8 @@ const defaultThumbnailSize = 256
 
 export const isOldEmote = (wearable: string): boolean => /^[a-z]+$/i.test(wearable)
 
-export const faceThumbnail: ValidateFn = validateAfterADR45(
-  async (components: ContentValidatorComponents, deployment: DeploymentToValidate) => {
+export function createFaceThumbnailValidateFn(components: ContentValidatorComponents) {
+  async function validateFn(deployment: DeploymentToValidate) {
     const errors: string[] = []
     const allAvatars: any[] = deployment.entity.metadata?.avatars ?? []
 
@@ -45,46 +52,48 @@ export const faceThumbnail: ValidateFn = validateAfterADR45(
     }
     return errors.length > 0 ? validationFailed(...errors) : OK
   }
-)
 
-export const wearableUrns: ValidateFn = validateAfterADR75(
-  async (components: ContentValidatorComponents, deployment: DeploymentToValidate) => {
-    const allAvatars: any[] = deployment.entity.metadata?.avatars ?? []
-    for (const avatar of allAvatars) {
-      for (const pointer of avatar.avatar.wearables) {
-        if (isOldEmote(pointer)) continue
+  return validateAfterADR45(validateFn)
+}
 
-        const parsed = await parseUrn(pointer)
-        if (!parsed)
-          return validationFailed(
-            `Each profile wearable pointer should be a urn, for example (urn:decentraland:{protocol}:collections-v2:{contract(0x[a-fA-F0-9]+)}:{name}). Invalid pointer: (${pointer})`
-          )
+export const wearableUrnsValidateFn = validateAfterADR75(async function validateFn(
+  deployment: DeploymentToValidate
+): Promise<ValidationResponse> {
+  const allAvatars: any[] = deployment.entity.metadata?.avatars ?? []
+  for (const avatar of allAvatars) {
+    for (const pointer of avatar.avatar.wearables) {
+      if (isOldEmote(pointer)) continue
+
+      const parsed = await parseUrn(pointer)
+      if (!parsed)
+        return validationFailed(
+          `Each profile wearable pointer should be a urn, for example (urn:decentraland:{protocol}:collections-v2:{contract(0x[a-fA-F0-9]+)}:{name}). Invalid pointer: (${pointer})`
+        )
+    }
+  }
+  return OK
+})
+
+export const emoteUrnsValidateFn = validateAfterADR74(async function validateFn(
+  deployment: DeploymentToValidate
+): Promise<ValidationResponse> {
+  const allAvatars = deployment.entity.metadata?.avatars ?? []
+  for (const avatar of allAvatars) {
+    const allEmotes = avatar.avatar.emotes ?? []
+    for (const { slot, urn } of allEmotes) {
+      if (isOldEmote(urn)) continue
+      const parsed = await parseUrn(urn)
+      if (!parsed)
+        return validationFailed(
+          `Each profile emote pointer should be a urn, for example (urn:decentraland:{protocol}:collections-v2:{contract(0x[a-fA-F0-9]+)}:{name}). Invalid pointer: (${urn})`
+        )
+      if (slot < 0 || slot > 9) {
+        return validationFailed(`The slot ${slot} of the emote ${urn} must be a number between 0 and 9 (inclusive).`)
       }
     }
-    return OK
   }
-)
-
-export const emoteUrns: ValidateFn = validateAfterADR74(
-  async (components: ContentValidatorComponents, deployment: DeploymentToValidate) => {
-    const allAvatars = deployment.entity.metadata?.avatars ?? []
-    for (const avatar of allAvatars) {
-      const allEmotes = avatar.avatar.emotes ?? []
-      for (const { slot, urn } of allEmotes) {
-        if (isOldEmote(urn)) continue
-        const parsed = await parseUrn(urn)
-        if (!parsed)
-          return validationFailed(
-            `Each profile emote pointer should be a urn, for example (urn:decentraland:{protocol}:collections-v2:{contract(0x[a-fA-F0-9]+)}:{name}). Invalid pointer: (${urn})`
-          )
-        if (slot < 0 || slot > 9) {
-          return validationFailed(`The slot ${slot} of the emote ${urn} must be a number between 0 and 9 (inclusive).`)
-        }
-      }
-    }
-    return OK
-  }
-)
+  return OK
+})
 
 function profileHasEmotes(deployment: DeploymentToValidate) {
   const allAvatars: Avatar[] = deployment.entity.metadata?.avatars ?? []
@@ -96,20 +105,27 @@ function profileHasEmotes(deployment: DeploymentToValidate) {
   return false
 }
 
-export const profileMustHaveEmotes: ValidateFn = validateAfterADR74(
-  async (components: ContentValidatorComponents, deployment: DeploymentToValidate) => {
-    if (!profileHasEmotes(deployment)) {
-      return validationFailed('Profile must have emotes after ADR 74.')
-    }
-    return OK
+export const profileMustHaveEmotesValidateFn = validateAfterADR74(async function validateFn(
+  deployment: DeploymentToValidate
+): Promise<ValidationResponse> {
+  if (!profileHasEmotes(deployment)) {
+    return validationFailed('Profile must have emotes after ADR 74.')
   }
-)
+  return OK
+})
 
-/**
- * Validate that given profile deployment includes the face256 file with the correct size
- * * @public
- */
-export const profile: ValidateFn = validateIfTypeMatches(
-  EntityType.PROFILE,
-  validateAll(faceThumbnail, wearableUrns, emoteUrns, profileMustHaveEmotes)
-)
+export function createProfileValidateFn(components: ContentValidatorComponents): ValidateFn {
+  /**
+   * Validate that given profile deployment includes the face256 file with the correct size
+   * * @public
+   */
+  return validateIfTypeMatches(
+    EntityType.PROFILE,
+    validateAll(
+      createFaceThumbnailValidateFn(components),
+      wearableUrnsValidateFn,
+      emoteUrnsValidateFn,
+      profileMustHaveEmotesValidateFn
+    )
+  )
+}
