+++ src/validations/scene.ts
@@ -1,25 +1,27 @@
 import { EntityType } from '@dcl/schemas'
-import { ContentValidatorComponents, DeploymentToValidate, OK, ValidateFn, validationFailed } from '../types'
+import { DeploymentToValidate, OK, ValidateFn, validationFailed, ValidationResponse } from '../types'
 import { validateAfterADR173, validateIfTypeMatches } from './validations'
 
-function sceneHasWorldConfiguration(deployment: DeploymentToValidate): boolean {
-  return deployment.entity.metadata?.worldConfiguration !== undefined
+function validateIfScene(validateFn: ValidateFn): ValidateFn {
+  return validateIfTypeMatches(EntityType.SCENE, validateFn)
 }
 
 /**
  * Validate that given scene deployment does not contain worldConfiguration section
  * @public
  */
-export const noWorldsConfiguration: ValidateFn = validateAfterADR173(
-  async (components: ContentValidatorComponents, deployment: DeploymentToValidate) => {
-    if (sceneHasWorldConfiguration(deployment)) {
+export const noWorldsConfigurationValidateFn = validateIfScene(
+  validateAfterADR173(async function validateFn(deployment: DeploymentToValidate): Promise<ValidationResponse> {
+    const sceneHasWorldConfiguration = deployment.entity.metadata?.worldConfiguration !== undefined
+    if (sceneHasWorldConfiguration) {
       return validationFailed('Scenes cannot have worldConfiguration section after ADR 173.')
     }
     return OK
-  }
+  })
 )
+
 /**
  * Validate that given scene deployment is valid
  * @public
  */
-export const scene: ValidateFn = validateIfTypeMatches(EntityType.SCENE, noWorldsConfiguration)
+export const sceneValidateFn = noWorldsConfigurationValidateFn
