+++ src/validations/subgraph-access-checker/emotes.ts
@@ -1,0 +1,8 @@
+import { SubgraphAccessCheckerComponents, ValidateFn } from '../../types'
+import { createItemValidateFn } from './items/items'
+
+export function createEmoteValidateFn(
+  components: Pick<SubgraphAccessCheckerComponents, 'externalCalls' | 'logs' | 'theGraphClient'>
+): ValidateFn {
+  return createItemValidateFn(components, ['blockchain-collection-v2-asset', 'blockchain-collection-third-party'])
+}
