+++ src/validations/subgraph-access-checker/items/off-chain-asset.ts
