+++ src/validations/subgraph-access-checker/items/third-party-asset.ts
@@ -3,7 +3,7 @@ import { keccak256Hash } from '@dcl/hashing'
 import { isThirdParty, MerkleProof, ThirdPartyProps } from '@dcl/schemas'
 import { BlockchainCollectionThirdParty } from '@dcl/urn-resolver'
 import { ILoggerComponent } from '@well-known-components/interfaces'
-import { ContentValidatorComponents, DeploymentToValidate, OK, validationFailed } from '../../../types'
+import { SubgraphAccessCheckerComponents, DeploymentToValidate, OK, validationFailed } from '../../../types'
 import { AssetValidation } from './items'
 
 function toHexBuffer(value: string): Buffer {
@@ -24,7 +24,7 @@ async function verifyHash<T>(
 ): Promise<boolean> {
   if (!MerkleProof.validate(metadata.merkleProof)) {
     logger.debug('Merkle proof is not valid', {
-      merkleProof: metadata.merkleProof
+      merkleProof: metadata.merkleProof,
     })
     return false
   }
@@ -34,7 +34,7 @@ async function verifyHash<T>(
   if (metadata.merkleProof.entityHash !== generatedCrcHash) {
     logger.debug(`The hash provided in the merkleProof doesn't match the one generated by the validator`, {
       generatedCrcHash,
-      entityHash: metadata.merkleProof.entityHash
+      entityHash: metadata.merkleProof.entityHash,
     })
     return false
   }
@@ -46,7 +46,7 @@ async function verifyHash<T>(
 }
 
 async function getMerkleRoot(
-  components: Pick<ContentValidatorComponents, 'externalCalls' | 'subGraphs'>,
+  components: Pick<SubgraphAccessCheckerComponents, 'externalCalls' | 'subGraphs'>,
   block: number,
   thirdPartyId: string
 ): Promise<string | undefined> {
@@ -61,14 +61,14 @@ async function getMerkleRoot(
     thirdParties: { root: string }[]
   }>(query, {
     id: thirdPartyId,
-    block
+    block,
   })
   if (!result.thirdParties || result.thirdParties.length < 1) return
   return result.thirdParties[0]?.root
 }
 
 async function verifyMerkleProofedEntity(
-  components: Pick<ContentValidatorComponents, 'externalCalls' | 'theGraphClient' | 'subGraphs'>,
+  components: Pick<SubgraphAccessCheckerComponents, 'externalCalls' | 'theGraphClient' | 'subGraphs'>,
   urn: BlockchainCollectionThirdParty,
   deployment: DeploymentToValidate,
   logger: ILoggerComponent.ILogger
@@ -110,7 +110,7 @@ async function verifyMerkleProofedEntity(
     logger.debug(`Merkle Tree based verifications failed:`, {
       merkleProof: JSON.stringify(metadata.merkleProof),
       merkleRoots: JSON.stringify(merkleRoots),
-      blocks: JSON.stringify([blockNumberAtDeployment, blockNumberFiveMinBeforeDeployment])
+      blocks: JSON.stringify([blockNumberAtDeployment, blockNumberFiveMinBeforeDeployment]),
     })
   }
 
@@ -119,7 +119,7 @@ async function verifyMerkleProofedEntity(
 
 export const thirdPartyAssetValidation: AssetValidation = {
   async validateAsset(
-    components: Pick<ContentValidatorComponents, 'externalCalls' | 'logs' | 'theGraphClient' | 'subGraphs'>,
+    components: Pick<SubgraphAccessCheckerComponents, 'externalCalls' | 'logs' | 'theGraphClient' | 'subGraphs'>,
     asset: BlockchainCollectionThirdParty,
     deployment: DeploymentToValidate
   ) {
@@ -133,5 +133,5 @@ export const thirdPartyAssetValidation: AssetValidation = {
   },
   canValidate(asset): asset is BlockchainCollectionThirdParty {
     return asset.type === 'blockchain-collection-third-party'
-  }
+  },
 }
